repo_results_2/run-20240312-153149_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-twitter-donnemartin-system-design-oop-whatsapp_0_20-runs_log/twitter/4/LLMatic/message.py
import datetime


class Message:
	def __init__(self, sender, receiver, content):
		self.sender = sender
		self.receiver = receiver
		self.content = content
		self.timestamp = datetime.datetime.now()
		self.blocked_users = []

	def send(self):
		if self.receiver in self.blocked_users:
			return 'User is blocked'
		else:
			return 'Message sent'

	def block_user(self, user):
		self.blocked_users.append(user)
		return 'User blocked'

	def unblock_user(self, user):
		if user in self.blocked_users:
			self.blocked_users.remove(user)
			return 'User unblocked'
		else:
			return 'User not found in blocked list'
