repo_results_2/run-20240312-153149_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-twitter-donnemartin-system-design-oop-whatsapp_0_20-runs_log/twitter/4/LLMatic/test_message.py
import unittest
from message import Message


class TestMessage(unittest.TestCase):
	def setUp(self):
		self.message = Message('User1', 'User2', 'Hello')

	def test_send(self):
		self.assertEqual(self.message.send(), 'Message sent')

	def test_block_user(self):
		self.assertEqual(self.message.block_user('User3'), 'User blocked')
		self.assertEqual(self.message.send(), 'Message sent')

	def test_unblock_user(self):
		self.message.block_user('User3')
		self.assertEqual(self.message.unblock_user('User3'), 'User unblocked')
		self.assertEqual(self.message.send(), 'Message sent')


if __name__ == '__main__':
	unittest.main()
